Move report and add ntp
Move report to its own file and add ntp/chronyd checks for the profile
